Renamed MapActivity to BrowseRequestRouteActivity, removed deprecated Browse activity. Moved fare calculator to RequestController from Route.
