+++ app/src/main/java/ca/ualberta/cs/unter/controller/RequestController.java
@@ -184,4 +184,10 @@ public class RequestController {
 
         updateRequest(request);
     }
+
+
+    public void calculateEstimatedFare(Request request) {
+        double fare = request.getDistance() * 0.50;
+		request.setEstimatedFare(fare);
+    }
 }
