+++ app/src/main/java/ca/ualberta/cs/unter/model/request/Request.java
@@ -44,12 +44,14 @@ import io.searchbox.core.SearchResult;
 /**
  * This is a class that contains all attributes a Request should have.
  */
-public abstract class Request implements FareCalculator{
+public abstract class Request {
     private String riderUserName;
     private String driverUserName;
     private ArrayList<String> driverList = new ArrayList<>();
 
     private Route route;
+
+	private double estimatedFare;
 
     private String requestDescription;
 
@@ -82,13 +84,13 @@ public abstract class Request implements FareCalculator{
      *
      * @param riderUserName         the rider user name
      * @param driverUserName        the driver user name
-     * @param estimatedFare         the estimated fare
      * @param route the path from pickup location to destination
      */
     public Request(String riderUserName, String driverUserName, Route route, double estimatedFare) {
         this.riderUserName = riderUserName;
         this.driverUserName = driverUserName;
         this.route = route;
+		this.estimatedFare = estimatedFare;
     }
 
     /**
@@ -96,12 +98,12 @@ public abstract class Request implements FareCalculator{
      *
      * @param riderUserName the rider user name
      * @param driverList the list of drivers username who accept the request
-     * @param estimatedFare the estimated fare
      */
     public Request(String riderUserName, ArrayList<String> driverList, Route route, double estimatedFare) {
         this.riderUserName = riderUserName;
         this.driverList = driverList;
         this.route = route;
+		this.estimatedFare = estimatedFare;
     }
 
     /**
@@ -429,15 +431,15 @@ public abstract class Request implements FareCalculator{
      * @return the estimated fare
      */
     public Double getEstimatedFare() {
-        return route.getEstimatedFare();
+        return estimatedFare;
     }
 
 	/**
-	 * Calculates the estimated fare for the route
-	 * @param distance
+	 * Sets the estimated fare.
+	 * @param fare
 	 */
-	public void calculateEstimatedFare(double distance) {
-		route.calculateFare(distance);
+	public void setEstimatedFare(double fare) {
+		estimatedFare = fare;
 	}
 
 	/**
